+++ changelog.md
@@ -78,5 +78,9 @@
 
 #1.0.17
 * Set node-sass version to 4.9.3
-* Fixed tour state change on initialize (It's possible to update steps even with active tour)
+* Fixed TourWizardService._tourState change on initialize (It's possible to update steps even with active tour)
+* Refactor
+
+#1.0.18
+* Fixed loosing currentStep reference when calling initialize
 * Refactor